+++ package.json
@@ -27,6 +27,7 @@
     "@eslint/js": "^9.25.0",
     "@storybook/addon-a11y": "^9.0.8",
     "@storybook/addon-docs": "^9.0.8",
+    "@storybook/addon-themes": "^9.0.8",
     "@storybook/addon-vitest": "^9.0.8",
     "@storybook/react": "^9.0.8",
     "@storybook/react-vite": "^9.0.8",
